splitting AuthModal to LoginForm and RegisterForm
