+++ packages/core/docs/changelog/5377.js
@@ -4,10 +4,16 @@ module.exports = {
   isBreaking: true,
   breakingChanges: [
     {
-      module: 'nuxt-theme',
+      module: 'nuxt-theme-module',
       before: 'LoginModal',
       after: 'AuthModal',
       comment: 'isLoginModal and toggleLoginModal was changed to isAuthModal and toggleAuthModal'
+    },
+    {
+      module: 'nuxt-theme-module',
+      before: 'AuthModal component',
+      after: 'LoginForm/RegisterForm component',
+      comment: 'Splitting AuthModal component to LoginForm and RegisterForm'
     }
   ],
   author: 'Łukasz Jędrasik',
