+++ packages/core/nuxt-theme-module/theme/layouts/default.vue
@@ -31,7 +31,7 @@ import AppFooter from '~/components/AppFooter.vue';
 import TopBar from '~/components/TopBar.vue';
 import CartSidebar from '~/components/CartSidebar.vue';
 import WishlistSidebar from '~/components/WishlistSidebar.vue';
-import AuthModal from '~/components/AuthModal.vue';
+import AuthModal from '~/components/Auth/AuthModal.vue';
 import LazyHydrate from 'vue-lazy-hydration';
 import Notification from '~/components/Notification';
